pyproject.toml
[project]
name = "financial-alpha-radar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.12",
    "google-genai>=1.70.0",
    "html2text>=2025.4.15",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.48",
    "yfinance>=1.2.0",
    "youtube-transcript-api>=1.2.4",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
